+++ src/utils/point-utils.js
@@ -1,14 +1,6 @@
 import dayjs from 'dayjs';
-import {POINTS_COUNT} from './const';
-
-// Функция из интернета по генерации случайного числа из диапазона
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {getRandomInteger} from '../utils/common';
+import {POINTS_COUNT} from '../const';
 
 const getDayFromDate = (date) => dayjs(date).format('D MMM');
 const getTimeFromDate = (date) => dayjs(date).format('hh:mm');
@@ -37,7 +29,6 @@ const generateArray = () => {
 };
 
 export {
-  getRandomInteger,
   generateArray,
   getDayFromDate,
   getTimeFromDate,
